requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
pillow>=8.0.0
timm>=0.5.0
tqdm>=4.62.0
scikit-learn>=1.0.0